+++ source/Nuke.SourceGenerators.Tests/StronglyTypedSolutionGeneratorTest.cs
@@ -3,10 +3,9 @@
 // https://github.com/nuke-build/nuke/blob/master/LICENSE
 
 using System;
-using System.IO;
 using System.Linq;
-using System.Reflection;
 using System.Threading.Tasks;
+using FluentAssertions;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Nuke.Common.ProjectModel;
@@ -19,7 +18,7 @@ namespace Nuke.SourceGenerators.Tests
     public class StronglyTypedSolutionGeneratorTest
     {
         [Fact]
-        public Task TestGlobalSolution()
+        public Task Test()
         {
             var inputCompilation = CreateCompilation(@"
 using Nuke.Common;
@@ -32,25 +31,61 @@ partial class Build : NukeBuild
 }");
 
             var generator = new StronglyTypedSolutionGenerator();
-            var driver = (GeneratorDriver) CSharpGeneratorDriver.Create(generator);
-            var result = driver.RunGeneratorsAndUpdateCompilation(inputCompilation, out _, out _).GetRunResult();
+            var driver = CSharpGeneratorDriver.Create(generator);
+            var result = driver.RunGenerators(inputCompilation);
+            return Verifier.Verify(result);
+        }
+
+        [Fact]
+        public void TestDisabled()
+        {
+            var inputCompilation = CreateCompilation(@"
+using Nuke.Common;
+using Nuke.Common.ProjectModel;
+
+partial class Build : NukeBuild
+{
+    [Solution(GenerateProjects = false)]
+    readonly Solution Solution;
+}");
+
+            var generator = new StronglyTypedSolutionGenerator();
+            var driver = CSharpGeneratorDriver.Create(generator);
+            var result = driver.RunGenerators(inputCompilation).GetRunResult();
 
             if (!result.Diagnostics.IsEmpty)
                 throw new Exception(string.Join(Environment.NewLine, result.Diagnostics.Select(x => x.GetMessage())));
-            return Verifier.Verify(result.GeneratedTrees.Single().GetRoot().ToFullString());
+            result.GeneratedTrees.Should().BeEmpty();
+        }
+
+        [Fact]
+        public void TestUnspecified()
+        {
+            var inputCompilation = CreateCompilation(@"
+using Nuke.Common;
+using Nuke.Common.ProjectModel;
+
+partial class Build : NukeBuild
+{
+    [Solution]
+    readonly Solution Solution;
+}");
+
+            var generator = new StronglyTypedSolutionGenerator();
+            var driver = CSharpGeneratorDriver.Create(generator);
+            var result = driver.RunGenerators(inputCompilation).GetRunResult();
+
+            if (!result.Diagnostics.IsEmpty)
+                throw new Exception(string.Join(Environment.NewLine, result.Diagnostics.Select(x => x.GetMessage())));
+            result.GeneratedTrees.Should().BeEmpty();
         }
 
         private static Compilation CreateCompilation(string source)
         {
             return CSharpCompilation.Create("compilation",
                 new[] { CSharpSyntaxTree.ParseText(source) },
-                new[]
-                {
-                    MetadataReference.CreateFromFile(typeof(object).GetTypeInfo().Assembly.Location),
-                    MetadataReference.CreateFromFile(Path.Combine(Path.GetDirectoryName(typeof(object).Assembly.Location)!, "netstandard.dll")),
-                    MetadataReference.CreateFromFile(Path.Combine(Path.GetDirectoryName(typeof(object).Assembly.Location)!, "System.Runtime.dll")),
-                    MetadataReference.CreateFromFile(typeof(SolutionAttribute).Assembly.Location)
-                },
+                new[] { MetadataReference.CreateFromFile(typeof(SolutionAttribute).Assembly.Location) }
+                    .Concat(Basic.Reference.Assemblies.NetStandard20.All),
                 new CSharpCompilationOptions(OutputKind.ConsoleApplication));
         }
     }